Extract LoadingDialog to separate module
