+++ main.py
@@ -75,21 +75,7 @@ METRIC_FIELD_ORDER = [
 ]
 
 
-class LoadingDialog(MDDialog):
-    """Simple dialog displaying a spinner while work is performed."""
-
-    def __init__(self, text: str = "Loading...", **kwargs):
-        box = MDBoxLayout(
-            orientation="vertical",
-            spacing="8dp",
-            size_hint_y=None,
-            height="72dp",
-        )
-        spinner = MDSpinner(size_hint=(None, None), size=("48dp", "48dp"))
-        spinner.pos_hint = {"center_x": 0.5}
-        box.add_widget(spinner)
-        box.add_widget(MDLabel(text=text, halign="center"))
-        super().__init__(type="custom", content_cls=box, **kwargs)
+from ui.loading_dialog import LoadingDialog
 
 
 class WorkoutActiveScreen(MDScreen):
